requirements.txt
flask
Pillow
tensorflow==2.15
numpy